requirements.txt
albucore==0.0.24
albumentations==2.0.7
annotated-types==0.7.0
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
filelock==3.13.1
fonttools==4.58.0
fsspec==2024.6.1
huggingface-hub==0.31.4
idna==3.10
imageio==2.37.0
Jinja2==3.1.4
joblib==1.5.0
kiwisolver==1.4.8
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
segmentation_models_pytorch==0.5.0
setuptools==70.2.0
simsimd==6.2.1
six==1.17.0
stringzilla==3.12.5
sympy==1.13.1
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.15
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
